ChafaSymbolMap: Fix crash when importing more than 32767 glyphs
The new limit is 2,147,483,646 glyphs. Any glyphs in excess of this
limit will be silently discarded.

Fixes #206 (GitHub).

+++ chafa/internal/chafa-private.h
@@ -84,7 +84,7 @@ struct ChafaSymbolMap
 
 typedef struct
 {
-    gint16 symbol_index;
+    gint symbol_index;
     guint8 hamming_distance;
     guint8 is_inverted;
 }
